Improve format text box, decimal and numeric inputs

+++ Joufflu.Inputs/Controls/TimeSpanPicker.xaml.cs
@@ -34,22 +34,23 @@ namespace Joufflu.Inputs.Controls
             if (Values.Count < 4)
                 return null;
 
-            // Groups are days / hours / minutes / seconds (see the Format above).
-            int? days = Values[0] as int?;
-            int? hours = Values[1] as int?;
-            int? minutes = Values[2] as int?;
-            int? seconds = Values[3] as int?;
+            // Groups are days / hours / minutes / seconds (see the Format above). A numeric group
+            // counts in long, and a boxed value only comes back out as the very type it went in as.
+            long? days = Values[0] as long?;
+            long? hours = Values[1] as long?;
+            long? minutes = Values[2] as long?;
+            long? seconds = Values[3] as long?;
 
             if (!days.HasValue || !hours.HasValue || !minutes.HasValue || !seconds.HasValue)
                 return null;
 
-            return new TimeSpan(days.Value, hours.Value, minutes.Value, seconds.Value);
+            return new TimeSpan((int)days.Value, (int)hours.Value, (int)minutes.Value, (int)seconds.Value);
         }
 
         public override List<object?> ConvertTo()
         {
             if (Value is TimeSpan date)
-                return new List<object?>() { date.Days, date.Hours, date.Minutes, date.Seconds };
+                return new List<object?>() { (long)date.Days, (long)date.Hours, (long)date.Minutes, (long)date.Seconds };
             else
                 return new List<object?>() { null, null, null, null };
         }
